% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_transform_coord.R
\name{affine_transform_coord}
\alias{affine_transform_coord}
\title{Apply an affine matrix to a coordinate pair}
\usage{
affine_transform_coord(df, coords = c("x", "y"), affine)
}
\description{
Perform an affine transformation on x and y.
}
