% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_overlay.R
\name{annotation_overlay}
\alias{annotation_overlay}
\title{Overlay an SLO with an en face computed image}
\usage{
annotation_overlay(plot_overlay, volume)
}
\description{
Overlays an SLO image with an en face image computed from the volume data.
This function facilitates combining two rasters with different fill
aesthetics. NOTE: I'm not sure if this will work with unregistered VOL files.
}
