% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_overlay_bscan.R
\name{annotation_overlay_bscan}
\alias{annotation_overlay_bscan}
\title{Overlay a B-scan with another image of the same dimensions}
\usage{
annotation_overlay_bscan(plot_overlay, volume)
}
\description{
Overlays an B-scan image with another image, such as segmentation surfaces,
which is the same dimensions as the original B-scan.
This function facilitates combining two rasters with different fill
aesthetics. NOTE: I'm not sure if this will work with unregistered VOL files.
}
