% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_overlay_enface.R
\name{annotation_overlay_enface}
\alias{annotation_overlay_enface}
\title{Overlay an en face image with another en face image}
\usage{
annotation_overlay_enface(plot_overlay, volume, xmin, xmax, ymin, ymax)
}
\description{
Overlays an en face image with another en face image computed from the volume
data.This function facilitates combining two rasters with different fill
aesthetics. NOTE: I'm not sure if this will work with unregistered VOL files.
}
