% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_volume.R
\name{apply_volume}
\alias{apply_volume}
\title{Apply a function across an array}
\usage{
apply_volume(volume, seg_volume, .fun, ...)
}
\description{
Applys a function across one array using the segmentation defined in a
second array.
}
