% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_grid_thickness.R
\name{calculate_grid_thickness}
\alias{calculate_grid_thickness}
\title{Compute thickness across OCT segmentations using a grid.}
\usage{
calculate_grid_thickness(
  segmentation,
  grid_center,
  grid,
  flip_region = "OD",
  laterality = grid_center[["scan_characteristics"]][["laterality"]][[1]],
  mask_undefined = TRUE
)
}
\description{
Compute regional thickness for segmentated OCT data.
}
