% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells_to_spatialpolygons.R
\name{cells_to_spatialpolygons}
\alias{cells_to_spatialpolygons}
\title{Convert grid regions to SpatialPolygons}
\usage{
cells_to_spatialpolygons(grid)
}
\description{
Convert grid regions to SpatialPolygons
}
