% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_layer_thickness.R
\name{compute_layer_thickness}
\alias{compute_layer_thickness}
\title{Compute layer thickness in voxels}
\usage{
compute_layer_thickness(seg, expand_surfaces = FALSE)
}
\description{
Compute per bscan layer thickness for every ascan.
}
