% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_thickness.R
\name{compute_thickness}
\alias{compute_thickness}
\title{Compute thickness of a segmentation array}
\usage{
compute_thickness(seg_array, na_undefined = FALSE, scale_value = FALSE)
}
\description{
Computes the thickness of a 2D segmentation array by subtracking the z values
of adjacent layers. NOTE: This might give unreliable results if layers
are allowed to intersect (i.e., after manual correction).
}
