% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_grid_to_slicer.R
\name{convert_grid_to_slicer}
\alias{convert_grid_to_slicer}
\title{Convert a 2D (x, y) grid and returns a 3D (x, y, z) segmentation array}
\usage{
convert_grid_to_slicer(grid, grid_center, dims, scale_x, scale_y)
}
\description{
Converts a 2D (x, y) grid and returns a 3D (x, y, z) segmentation array.
}
