% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seg_array.R
\name{create_seg_array}
\alias{create_seg_array}
\title{Convert segmentation object into a segmentation array}
\usage{
create_seg_array(segmentation)
}
\description{
Converts a segmentation object from `heyexr` into a segmentation array.
This function works with IRA/OCTExplorer segmenation coordinate definition.
}
