% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_slicer_range.R
\name{create_slicer_range}
\alias{create_slicer_range}
\title{Create a binary slicer from a segmentation array}
\usage{
create_slicer_range(seg_volume, from, to)
}
\description{
These two functions create binary slicers from segmentation arrays by either
specifying a number range of values or by specifying values directly.
}
