% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_slicer_subset.R
\name{create_slicer_subset}
\alias{create_slicer_subset}
\title{Create a binary slicer from a segmentation array}
\usage{
create_slicer_subset(seg_volume, values)
}
\description{
These two functions create binary slicers from segmentation arrays by either
specifying a number range of values or by specifying values directly.
}
