% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.calc_y}
\alias{.calc_y}
\title{Given an rows, columns, and radius, determine the grid of hexagon centers}
\usage{
.calc_y(i, j, s, y0 = 0)
}
\description{
Code from:
http://hacksociety.net/Thread-Tutorial-Creating-a-hexagonal-grid-for-games-C
}
