% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_first.R
\name{.repeat_first}
\alias{.repeat_first}
\title{Repeat the first row at the end of the data.frame.}
\usage{
.repeat_first(df)
}
\arguments{
\item{df}{A data.frame}
}
\value{
A data.frame with the first row repeated at the end.
}
\description{
Repeat the first row at the end of the data.frame.
}
\keyword{internal}
