% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_surfaces.R
\name{expand_surfaces}
\alias{expand_surfaces}
\title{Add pseudo-surfaces to the layer table.}
\usage{
expand_surfaces(seg)
}
\description{
This function adds two new surfaces. They represent the inner-most and outer-
most edges of the volume. These surfaces are not segmented by the
Iowa Reference Algorithms.
}
