% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_surfaces_to_volume.R
\name{expand_surfaces_to_volume}
\alias{expand_surfaces_to_volume}
\title{Expand segmentation surfaces into a segmentation volume}
\usage{
expand_surfaces_to_volume(surface_array, vol_dim)
}
\description{
Expands segmentation surfaces into a segmentation volume. Surfaces IDs are
imputed from the inner inner surface of a layer. NOTE: This assumes surfaces
IDs start at 1. Creates artificial surfaces at the top and bottoms of the
volume.
}
