% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_replace.R
\name{find_replace}
\alias{find_replace}
\title{Replace elements based on a function}
\usage{
find_replace(.x, replacement, find_fun, ...)
}
\description{
Replaces elements based on a function which returns a logical condition.
}
