% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sppoints_in_sppolygons.R
\name{find_sppoints_in_sppolygons}
\alias{find_sppoints_in_sppolygons}
\title{Find points in regions}
\usage{
find_sppoints_in_sppolygons(pts, plygns, pts_attributes = NULL)
}
\description{
Determine which points appear in which regions.
}
