% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_array.R
\name{flip_array}
\alias{flip_array}
\title{Flip an array along an axis.}
\usage{
flip_array(x, flip_along)
}
\description{
Flips an array along a specified axis. Idea from
https://everydropr.wordpress.com/2012/10/16/a-simple-way-to-flip-a-matrix-or-an-array/
NOTE: I'm not sure how well this function behaves.
}
