% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_correction_array.R
\name{gamma_correction_array}
\alias{gamma_correction_array}
\title{Apply gamma correction on an array of data}
\usage{
gamma_correction_array(a, gamma = 1/4)
}
\description{
Applys a gamma correction to an array. Default is the value suggested by
Heidelberg for rendering B-scans in VOL files.
}
