% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_circle.R
\name{generate_circle}
\alias{generate_circle}
\title{Create a circle}
\usage{
generate_circle(center = c(0, 0), radius = 1, npoints = 100)
}
\description{
Code based on:
http://gastonsanchez.com/blog/how-to/2012/06/17/PCA-in-R.html
}
