% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_grid_hexagons.R
\name{generate_grid_hexagons}
\alias{generate_grid_hexagons}
\title{Generate a grid of regular hexagons}
\usage{
generate_grid_hexagons(nrow, ncol, radius)
}
\description{
Generate a grid of regular hexagons
}
