% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_grid_rectangles.R
\name{generate_grid_rectangles}
\alias{generate_grid_rectangles}
\title{Generate a grid of rectangles}
\usage{
generate_grid_rectangles(center_x, center_y, nrow, ncol, width, height = width)
}
\description{
Generates a grid of rectangles around a center point
}
