% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hexagon_centers.R
\name{generate_hexagon_centers}
\alias{generate_hexagon_centers}
\title{Generate hexagon centers}
\usage{
generate_hexagon_centers(nrow, ncol, radius, x0 = 0, y0 = 0)
}
\description{
Generate the center points associated with a grid of hexagonal cells.
}
