% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hexagon_nodes.R
\name{generate_hexagon_nodes}
\alias{generate_hexagon_nodes}
\title{Generate hexagon nodes}
\usage{
generate_hexagon_nodes(center_x, center_y, radius = 1, rotation = 0)
}
\description{
Give a center point, radius, and rotation, return the line segments
for a hexagon
}
