% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hexagon_segments.R
\name{generate_hexagon_segments}
\alias{generate_hexagon_segments}
\title{Generate hexagon segments}
\usage{
generate_hexagon_segments(center_x, center_y, radius = 1, rotation = 0)
}
\description{
Generates segments corresponding to hexagonal shapes.
}
