% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_array_center.R
\name{get_array_center}
\alias{get_array_center}
\title{Compute the center coordinate of an array}
\usage{
get_array_center(a)
}
\description{
Computes the center coordiante of an array. Useful for setting the origin
in `rgl` 3D plots.
}
