% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enface_aspect.R
\name{get_enface_aspect}
\alias{get_enface_aspect}
\title{Compute the proper aspect ratio of the en face image}
\usage{
get_enface_aspect(volume)
}
\description{
Computes the aspect ratio of an en face image from a volume object. Useful
for rendering en face images in `ggplot2`.
}
