% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legend.R
\name{get_legend}
\alias{get_legend}
\title{Extract a legend from a `ggplot` object}
\usage{
get_legend(p)
}
\description{
Extracts the legend from a `ggplot` object.
Legend function adapted from
http://stackoverflow.com/questions/16963137/ggplot2-multiple-plots-with-different-variables-in-a-single-row-single-groupin
}
