% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slo_center.R
\name{get_slo_center}
\alias{get_slo_center}
\title{Determine the center coordinates of a `grid center` object in SLO space}
\usage{
get_slo_center(volume, grid_center)
}
\description{
Determines the center coordinate of a `grid center` object in SLO space.
}
