% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_undefined.R
\name{get_undefined}
\alias{get_undefined}
\title{Get the undefined values from a `segmentation` object}
\usage{
get_undefined(segmentation)
}
\description{
Gets the undefined A-scans from a `segmentation` object.
}
