% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_undefined_array.R
\name{get_undefined_array}
\alias{get_undefined_array}
\title{Get the undefined values from a `segmentation` object as an array}
\usage{
get_undefined_array(segmentation)
}
\description{
Returns the undefined A-scans from a `segmentation` object as an array.
}
