% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_undefined_matrix.R
\name{get_undefined_matrix}
\alias{get_undefined_matrix}
\title{Get the undefined values from a `segmentation` object as a matrix}
\usage{
get_undefined_matrix(segmentation)
}
\description{
Returns the undefined A-scans from a `segmentation` object as a matrix.
}
