% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_volume_angle_in_slo.R
\name{get_volume_angle_in_slo}
\alias{get_volume_angle_in_slo}
\title{Return angle of B-scans in radians}
\usage{
get_volume_angle_in_slo(volume)
}
\description{
Returns the angle of B-scans in a VOL file.
}
\keyword{internal}
