% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_volume_bbox_center_in_slo.R
\name{get_volume_bbox_center_in_slo}
\alias{get_volume_bbox_center_in_slo}
\title{Return the center of the bounding box of volume data in an SLO}
\usage{
get_volume_bbox_center_in_slo(volume)
}
\description{
Returns the center of the bounding box of volume data in an SLO. Assumes
Heidelberg VOL file data.
}
