% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_volume_bbox_in_slo.R
\name{get_volume_bbox_in_slo}
\alias{get_volume_bbox_in_slo}
\title{TASK: Test this slope function on VOL files with known rotated volume scans.}
\usage{
get_volume_bbox_in_slo(volume)
}
\description{
Return the volume bounding box in pixel coordinates, adjusted for the voxel
"size."
}
\keyword{internal}
