% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iowa_segmentation_to_array.R
\name{iowa_segmentation_to_array}
\alias{iowa_segmentation_to_array}
\title{Convert Iowa segmentation data.frame to array}
\usage{
iowa_segmentation_to_array(segmentation, mask_undefined = FALSE)
}
\description{
Convert Iowa segmentation data.frame to an array.
Produces an array in the shape of:
dim(# ascans, # bscans, # surface_id)
}
