% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_points.R
\name{make_points}
\alias{make_points}
\title{Convert x and y columns in a data.frame to SpatialPoints}
\usage{
make_points(df, x = "x", y = "y")
}
\description{
Takes a data.frame containing x and y coordinates and returns a SpatialPoints object
}
