% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_volume.R
\name{map_volume}
\alias{map_volume}
\title{Map a function across a volume using a segmentation volume}
\usage{
map_volume(volume, seg_volume, .fun, ...)
}
\description{
Maps a function across a volume using a segmentation volume. Returns a list.
}
