% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_volume_xy.R
\name{map_volume_xy}
\alias{map_volume_xy}
\title{Map a function across a volume using a segmentation volume}
\usage{
map_volume_xy(volume, seg_volume, .fun, ...)
}
\description{
Maps a function across a volume using a segmentation volume. Returns a list
of matrices.
}
