% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_matrix.R
\name{normalize_matrix}
\alias{normalize_matrix}
\title{Normalize values in a matrix}
\usage{
normalize_matrix(m)
}
\description{
Normalizes values in a matrix to lie between 0-1. NOTE: Perhaps this should
be termed `rescale_matrix`.
}
