% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over_tibble.R
\name{over_tibble}
\alias{over_tibble}
\title{A wrapper for sp::over that allows one to bind data to the resulting tibble}
\usage{
over_tibble(x, y, data = NULL)
}
\description{
A wrapper for sp::over that allows one to bind data to the resulting tibble
}
