% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangle_to_polygon.R
\name{rectangle_to_polygon}
\alias{rectangle_to_polygon}
\title{Convert rectangular min and max points to a polygon}
\usage{
rectangle_to_polygon(xmin, xmax, ymin, ymax)
}
\description{
Convert two points representing a rectangle to four points representing a
polygon.
}
