% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments_to_polygons.R
\name{segments_to_polygons}
\alias{segments_to_polygons}
\title{Convert segments to polygons}
\usage{
segments_to_polygons(segments_by_group, group, data, x = "x", y = "y")
}
\description{
Takes a data.frame of segments grouped by polygon and returns a SpatialPolygonsDataFrame
}
