% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_correction_array.R
\name{spline_correction_array}
\alias{spline_correction_array}
\title{Apply spline-based correction on an array containing Heidelberg B-scans}
\usage{
spline_correction_array(a)
}
\description{
Applys a spline-based correction to an array. Correction is based modeled on
the relationship between uncorrected values and the PNGs exported from
Heidelberg Explorer.
}
