% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_region_thickness.R
\name{summarize_cell_thickness}
\alias{summarize_cell_thickness}
\title{Summarize layer thickness across defined regions}
\usage{
summarize_cell_thickness(pts_cell, cell_centers)
}
\description{
Summarize segmented retinal layer thickness across regions defined by a grid.
}
