% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_to_spatialpoints.R
\name{tibble_to_spatialpoints}
\alias{tibble_to_spatialpoints}
\title{Convert a data.frame to a SpatialPoints object}
\usage{
tibble_to_spatialpoints(tbl)
}
\description{
Convert a data.frame to a SpatialPoints object
}
