% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_regions.R
\name{unique_regions}
\alias{unique_regions}
\title{Identify the unique region IDs in a segmentation volume}
\usage{
unique_regions(seg_volume)
}
\description{
Identifies the unique region IDs in a segmetnation volume.
}
